chore: refactored signup wizard provider trigger logic
