+++ packages/tradershub/src/AppContent.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import SignupWizard from './flows/RealAccountSIgnup/SignupWizard';
 import useRegulationFlags from './hooks/useRegulationFlags';
 import { EUDisclaimerMessage, useUIContext } from './components';
 import { Router } from './routes';
@@ -16,7 +15,6 @@ const AppContent = () => {
                 <Router />
             </div>
             {isEU && <EUDisclaimerMessage />}
-            <SignupWizard />
         </div>
     );
 };
