+++ packages/tradershub/src/providers/SignupWizardProvider/types.ts
@@ -7,8 +7,6 @@ export type TSignupWizardContext = {
     currentStep: number;
     dispatch: React.Dispatch<TActions>;
     helpers: Helpers;
-    isWizardOpen: boolean;
-    setIsWizardOpen: React.Dispatch<React.SetStateAction<boolean>>;
     state: TState;
 };
